requirements.txt
streamlit
openai
PyPDF2
requests
beautifulsoup4
python-dotenv
selenium
chromedriver